压缩图片

+++ app/v1/eblock/model/unit.py
@@ -284,27 +284,33 @@ class Unit(BaseModel):
                 # 其他情况下复制到公共读的目录 比如txt文件等 其他unit需要用到
                 shutil.copyfile(os.path.join(self.unit_work_path, file), target_read_path)
 
-            # 针对结果为0的unit，进行rds图片的压缩，为了减小体积
-            if self.detail.get("result") == 0:
-                if os.path.exists(target_path):
-                    # 针对png图片进行压缩
-                    if target_path.endswith('png'):
-                        # 先缩放
-                        origin_pic = cv2.imread(target_path, cv2.IMREAD_COLOR)
-                        if origin_pic is not None:
-                            origin_size = origin_pic.shape
-                            new_size = (
-                                int(origin_size[1] * PICTURE_COMPRESS_RATIO),
-                                int(origin_size[0] * PICTURE_COMPRESS_RATIO))
-                            compress_pic = cv2.resize(origin_pic, new_size)
-                            cv2.imwrite(target_path, compress_pic)
-                            # 后压缩
-                            self.pngquant_compress(target_path)
+            # 针对png图片进行压缩
+            if os.path.exists(target_path) and target_path.endswith('png'):
+                # 针对结果为0的unit，进行rds图片的压缩，为了减小体积
+                if self.detail.get("result") == 0 or self.can_compress(file):
+                    # 先缩放
+                    origin_pic = cv2.imread(target_path, cv2.IMREAD_COLOR)
+                    if origin_pic is not None:
+                        origin_size = origin_pic.shape
+                        new_size = (
+                            int(origin_size[1] * PICTURE_COMPRESS_RATIO),
+                            int(origin_size[0] * PICTURE_COMPRESS_RATIO))
+                        compress_pic = cv2.resize(origin_pic, new_size)
+                        cv2.imwrite(target_path, compress_pic)
+                        # 后压缩
+                        self.pngquant_compress(target_path)
 
     @staticmethod
     def pngquant_compress(fp):
         command = f'pngquant {fp} -f --quality 100-100'
         subprocess.run(command, shell=True)
+
+    def can_compress(self, target_filename):
+        for filepath in self.detail.get('not_compress_png_list', []):
+            _, filename = os.path.split(filepath)
+            if filename == target_filename:
+                return False
+        return True
 
     @staticmethod
     def remove_duplicate_pic(path):
